backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-k"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
description = "AGENT-K multi-agent Kaggle competition framework"
readme = "README.md"
license = "MIT"
dependencies = [
    "pydantic-ai>=0.2.0",
    "httpx>=0.27.0",
    "logfire>=0.44.0",
    "pydantic-graph>=0.2.0",
    "opentelemetry-api>=1.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_k"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agent_k"]

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
